Refactor large modules into focused utilities
