+++ src/components/reports/ConfirmationModal.jsx
@@ -1,0 +1,35 @@
+import React from 'react';
+
+const ConfirmationModal = ({ title, description, confirmLabel, cancelLabel = 'Cancelar', onConfirm, onCancel }) => (
+  <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/20 backdrop-blur-sm animate-fade-in">
+    <div className="bg-white rounded-2xl p-6 max-w-md w-full mx-4 shadow-2xl border border-secondary-200">
+      <div className="flex items-start gap-4 mb-4">
+        <div className="w-12 h-12 rounded-full bg-red-100 flex items-center justify-center flex-shrink-0">
+          <svg className="w-6 h-6 text-red-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z" />
+          </svg>
+        </div>
+        <div className="flex-1">
+          <h3 className="text-lg font-bold text-secondary-900 mb-2">{title}</h3>
+          <p className="text-sm text-secondary-600">{description}</p>
+        </div>
+      </div>
+      <div className="flex items-center gap-3 justify-end">
+        <button
+          onClick={onCancel}
+          className="px-4 py-2 rounded-xl text-sm font-semibold text-secondary-700 hover:bg-secondary-100 transition-colors"
+        >
+          {cancelLabel}
+        </button>
+        <button
+          onClick={onConfirm}
+          className="px-4 py-2 rounded-xl text-sm font-semibold bg-red-600 text-white hover:bg-red-700 transition-colors"
+        >
+          {confirmLabel}
+        </button>
+      </div>
+    </div>
+  </div>
+);
+
+export default ConfirmationModal;
