+++ src/context/AppContext.jsx
@@ -3,7 +3,6 @@ import React, { createContext, useState, useEffect, useMemo, useContext, useRef 
 import { callAiApi } from '../lib/apiService';
 import { PROMPT_FLOW_ANALYSIS_FROM_IMAGES, PROMPT_REFINE_FLOW_ANALYSIS_FROM_IMAGES_AND_CONTEXT } from '../lib/prompts';
 import { validateAndCleanImages, compressImageIfNeeded } from '../lib/imageService';
-
 import {
     loadReports as loadReportsFromDB,
     saveReport as saveReportToDB,
@@ -19,6 +18,7 @@ import {
     createUserStory,
     deleteUserStory
 } from '../lib/databaseService';
+import { cleanReportData, extractPasoFieldsFromText } from './utils/reportCleaning';
 import { supabase } from '../lib/supabaseClient';
 
 const AppContext = createContext();
@@ -93,324 +93,6 @@ export const AppProvider = ({ children }) => {
             activeMainMenu: mainMenu || prev.activeMainMenu,
             ...(subMenu && { activeSubMenu: subMenu })
         }));
-    };
-
-    const extractPasoFieldsFromText = (text = '') => {
-        if (!text) return {};
-
-        const normalized = text.replace(/\r/g, '');
-        const extract = (label) => {
-            const regex = new RegExp(`${label}\s*[:\-]\s*([^|\n]+)`, 'i');
-            const match = normalized.match(regex);
-            return match ? match[1].trim() : null;
-        };
-
-        return {
-            datoEntrada: extract('(?:dato(?:s)? de entrada|dato(?:s)? ancla|input data)') || undefined,
-            resultadoEsperado: extract('resultado esperado(?: del paso)?') || extract('validaci[oó]n esperada') || undefined,
-            resultadoObtenido: extract('resultado obtenido(?: del paso)?') || extract('observaci[oó]n') || undefined
-        };
-    };
-
-    // Function to clean and normalize AI response fields
-    const cleanReportData = (reportData, images = []) => {
-        if (!reportData) return reportData;
-
-        let cleaned = { ...reportData };
-
-        // --- NORMALIZACIÓN DE CLAVES (Inglés -> Español y Variantes) ---
-
-        // 1. ID y Título
-        if (!cleaned.id_caso) cleaned.id_caso = cleaned.test_case_id || cleaned.id || 'Generado';
-
-        // Detectar IDs técnicos inválidos y simplificarlos
-        // Detectar IDs técnicos inválidos y simplificarlos, pero mantener unicidad
-        if (typeof cleaned.id_caso === 'string') {
-            // Si es 'Generado' o vacío, asignar uno temporal único
-            if (cleaned.id_caso === 'Generado' || !cleaned.id_caso) {
-                cleaned.id_caso = `GEN-${Date.now().toString().slice(-4)}`;
-            }
-            // Si el ID es muy largo o complejo, intentar limpiarlo pero sin perder identidad
-            else if (cleaned.id_caso.length > 15) {
-                // Intentar extraer la parte más significativa (ej: TC-123 de HU-45-TC-123)
-                const parts = cleaned.id_caso.split(/[-_]/);
-                if (parts.length > 1) {
-                    // Tomar la última parte si parece un número o código corto
-                    const lastPart = parts[parts.length - 1];
-                    if (lastPart.length < 6) {
-                        cleaned.id_caso = lastPart;
-                    }
-                }
-            }
-        }
-
-        if (!cleaned.escenario_prueba) {
-            cleaned.escenario_prueba = cleaned.title || cleaned.titulo || cleaned.nombre_escenario || 'Definir nombre del escenario';
-        }
-
-        // Detectar nombres genéricos o técnicos y reemplazarlos
-        const genericNames = ['Caso de Prueba', 'Flujo de Usuario', 'Test Case', 'Escenario de Prueba'];
-        const technicalPrefixes = ['E2E-', 'TC-', 'Refinamiento_', 'UI-'];
-
-        if (typeof cleaned.escenario_prueba === 'string') {
-            const trimmed = cleaned.escenario_prueba.trim();
-
-            // Si es un nombre genérico exacto
-            if (genericNames.includes(trimmed)) {
-                cleaned.escenario_prueba = 'Definir nombre del escenario';
-            }
-            // Si empieza con prefijos técnicos, intentar extraer la parte descriptiva
-            else if (technicalPrefixes.some(prefix => trimmed.startsWith(prefix))) {
-                // Intentar limpiar el nombre técnico
-                let cleanedName = trimmed;
-                technicalPrefixes.forEach(prefix => {
-                    cleanedName = cleanedName.replace(prefix, '');
-                });
-                // Reemplazar guiones/underscores con espacios y capitalizar
-                cleanedName = cleanedName.replace(/[-_]/g, ' ').trim();
-                cleaned.escenario_prueba = cleanedName || 'Definir nombre del escenario';
-            }
-        }
-
-        // 2. Precondiciones (Manejo robusto de Arrays y Strings JSON)
-        let rawPreconditions = cleaned.precondiciones || cleaned.pre_conditions || cleaned.preconditions;
-        if (rawPreconditions) {
-            // Detectar valores inválidos
-            const invalidValues = ['-', 'N/A', 'n/a', 'NA', 'null', ''];
-            const trimmed = typeof rawPreconditions === 'string' ? rawPreconditions.trim() : rawPreconditions;
-
-            if (invalidValues.includes(trimmed)) {
-                cleaned.precondiciones = 'Ninguna precondición específica';
-            } else if (Array.isArray(rawPreconditions)) {
-                cleaned.precondiciones = rawPreconditions.map(p => `- ${p}`).join('\n');
-            } else if (typeof rawPreconditions === 'string') {
-                // Intentar detectar si es un string con formato JSON de array '["a", "b"]'
-                if (trimmed.startsWith('[') && trimmed.endsWith(']')) {
-                    try {
-                        const parsed = JSON.parse(trimmed);
-                        if (Array.isArray(parsed)) {
-                            cleaned.precondiciones = parsed.map(p => `- ${p}`).join('\n');
-                        } else {
-                            cleaned.precondiciones = rawPreconditions;
-                        }
-                    } catch (e) {
-                        cleaned.precondiciones = rawPreconditions; // Si falla, dejar como está
-                    }
-                } else {
-                    cleaned.precondiciones = rawPreconditions;
-                }
-            }
-        } else {
-            cleaned.precondiciones = 'Ninguna precondición específica';
-        }
-
-        // 3. Resultados (Esperado y Obtenido)
-        const invalidValues = ['-', 'N/A', 'n/a', 'NA', 'null', ''];
-
-        if (!cleaned.resultado_esperado) {
-            // Intentar buscar en otros campos (incluyendo variantes globales)
-            cleaned.resultado_esperado = cleaned.expected_result ||
-                cleaned.resultado_esperado_general ||
-                cleaned.resultado_esperado_flujo ||
-                cleaned.resultado_esperado_global;
-
-            // Si sigue faltando, usar poscondiciones como fallback (ya que describen el éxito del flujo)
-            if (!cleaned.resultado_esperado) {
-                let posConds = cleaned.poscondiciones || cleaned.post_conditions || cleaned.postconditions || cleaned.pos_conditions;
-                if (posConds) {
-                    if (Array.isArray(posConds)) {
-                        cleaned.resultado_esperado = posConds.map(p => `- ${p}`).join('\n');
-                    } else {
-                        cleaned.resultado_esperado = posConds;
-                    }
-                }
-            }
-
-            // Default final
-            if (!cleaned.resultado_esperado) cleaned.resultado_esperado = 'Definir criterio de éxito esperado';
-        }
-
-        // Detectar valores inválidos en resultado_esperado
-        if (typeof cleaned.resultado_esperado === 'string' && invalidValues.includes(cleaned.resultado_esperado.trim())) {
-            cleaned.resultado_esperado = 'Definir criterio de éxito esperado';
-        }
-
-        if (!cleaned.resultado_obtenido) {
-            cleaned.resultado_obtenido = cleaned.actual_result ||
-                cleaned.resultado_actual ||
-                cleaned.conclusion ||
-                cleaned.estado_final ||
-                cleaned.resultado_obtenido_global ||
-                'Pendiente de ejecución';
-        }
-
-        // Detectar valores inválidos en resultado_obtenido
-        if (typeof cleaned.resultado_obtenido === 'string' && invalidValues.includes(cleaned.resultado_obtenido.trim())) {
-            cleaned.resultado_obtenido = 'Pendiente de ejecución';
-        }
-
-        // POST-PROCESAMIENTO INTELIGENTE: Si hay imágenes pero el resultado es "Pendiente", inferir del esperado
-        if (images.length > 0 && cleaned.resultado_obtenido === 'Pendiente de ejecución' && cleaned.resultado_esperado && cleaned.resultado_esperado !== 'Definir criterio de éxito esperado') {
-            // Convertir el resultado esperado en observado (cambiar "debería" por "se observa")
-            cleaned.resultado_obtenido = cleaned.resultado_esperado
-                .replace(/debería/gi, 'se observa que')
-                .replace(/debe/gi, 'se visualiza')
-                .replace(/Se visualiza/gi, 'Se visualiza correctamente');
-        }
-
-        // POST-PROCESAMIENTO DE ESTADO GENERAL: Comparación semántica inteligente
-        if (!cleaned.estado_general || cleaned.estado_general === 'Pendiente') {
-            // Si ambos resultados existen y no son placeholders
-            if (cleaned.resultado_esperado &&
-                cleaned.resultado_obtenido &&
-                cleaned.resultado_esperado !== 'Definir criterio de éxito esperado' &&
-                cleaned.resultado_obtenido !== 'Pendiente de ejecución') {
-
-                // Normalizar para comparación (quitar puntuación, minúsculas, espacios extra)
-                const normalizar = (texto) => texto.toLowerCase().replace(/[.,;:]/g, '').replace(/\s+/g, ' ').trim();
-                const esperadoNorm = normalizar(cleaned.resultado_esperado);
-                const obtenidoNorm = normalizar(cleaned.resultado_obtenido);
-
-                // Verificar si contienen palabras clave de error
-                const tieneError = /error|fallo|fallido|incorrecto|no se pudo|denegado/i.test(obtenidoNorm);
-
-                // Si no hay errores y tienen similitud semántica (más del 40% de palabras en común)
-                if (!tieneError) {
-                    const palabrasEsperado = esperadoNorm.split(' ').filter(p => p.length > 3);
-                    const palabrasObtenido = obtenidoNorm.split(' ').filter(p => p.length > 3);
-                    const palabrasComunes = palabrasEsperado.filter(p => palabrasObtenido.includes(p));
-                    const similitud = palabrasComunes.length / Math.max(palabrasEsperado.length, 1);
-
-                    if (similitud > 0.4) {
-                        cleaned.estado_general = 'Exitoso';
-                    } else {
-                        cleaned.estado_general = 'Pendiente';
-                    }
-                } else {
-                    cleaned.estado_general = 'Fallido';
-                }
-            } else if (cleaned.resultado_obtenido === 'Pendiente de ejecución') {
-                cleaned.estado_general = 'Pendiente';
-            }
-        }
-
-        // 4. Mapeo de pasos (steps -> pasos) y Asignación de Imágenes
-        let rawSteps = cleaned.pasos || cleaned.steps || cleaned.Pasos_Analizados;
-
-        // DEBUG: Ver qué estructura tienen los pasos
-        if (rawSteps && Array.isArray(rawSteps) && rawSteps.length > 0) {
-            console.log('DEBUG - Estructura del primer paso:', rawSteps[0]);
-            console.log('DEBUG - Tipo del primer paso:', typeof rawSteps[0]);
-            console.log('DEBUG - Claves disponibles:', Object.keys(rawSteps[0]));
-        }
-
-        if (rawSteps && Array.isArray(rawSteps)) {
-            // CONVERSIÓN: Si los pasos son strings en lugar de objetos, convertirlos
-            if (rawSteps.length > 0 && typeof rawSteps[0] === 'string') {
-                console.warn('⚠️ La IA generó pasos como strings. Convirtiendo a objetos...');
-                rawSteps = rawSteps.map((stepText, index) => ({
-                    numero_paso: index + 1,
-                    descripcion: stepText,
-                    imagen_referencia: index < images.length ? `Evidencia ${index + 1}` : (images.length > 0 ? `Evidencia ${images.length}` : 'N/A')
-                }));
-            }
-
-            console.log('🔍 Procesando pasos. Total:', rawSteps.length);
-            console.log('🔍 Primer paso raw:', rawSteps[0]);
-
-            cleaned.Pasos_Analizados = rawSteps.map((step, index) => {
-                // Determinar la referencia de imagen
-                let imgRef = step.imagen_referencia || step.image_ref || step.evidencia;
-
-                // Si no tiene referencia explícita, asignar por índice secuencial
-                if (!imgRef || imgRef === 'N/A') {
-                    if (index < images.length) {
-                        imgRef = `Evidencia ${index + 1}`;
-                    } else if (images.length > 0) {
-                        // Fallback: usar la última imagen disponible para pasos adicionales
-                        imgRef = `Evidencia ${images.length}`;
-                    } else {
-                        imgRef = 'N/A';
-                    }
-                }
-
-                return {
-                    numero_paso: step.numero_paso || step.step_number || step.number || step.id_paso || step.orden || (index + 1),
-                    descripcion_accion_observada: step.descripcion_accion_observada || step.descripcion || step.description || step.action || step.accion || step.texto || step.text || 'Sin descripción',
-                    imagen_referencia: imgRef
-                };
-            });
-
-            console.log('✅ Pasos procesados. Primer paso final:', cleaned.Pasos_Analizados[0]);
-        }
-
-        // Mapeo de claves legacy para compatibilidad UI
-        cleaned.Nombre_del_Escenario = cleaned.escenario_prueba;
-        cleaned.Resultado_Esperado_General_Flujo = cleaned.resultado_esperado;
-        cleaned.Conclusion_General_Flujo = cleaned.resultado_obtenido;
-        // ---------------------------------------------------
-
-        // Clean Resultado_Esperado_General_Flujo field
-        if (cleaned.Resultado_Esperado_General_Flujo) {
-            let cleanedField = cleaned.Resultado_Esperado_General_Flujo;
-            // Remove various patterns of redundant text
-            const patterns = [
-                /^Resultado Esperado General del Flujo:\s*/i,
-                /^Resultado Esperado General:\s*/i,
-                /^Resultado Esperado:\s*/i
-            ];
-
-            for (const pattern of patterns) {
-                cleanedField = cleanedField.replace(pattern, '');
-            }
-
-            cleaned.Resultado_Esperado_General_Flujo = cleanedField.trim();
-        }
-
-        // Clean resultado_obtenido_paso_y_estado fields in Pasos_Analizados
-        if (cleaned.Pasos_Analizados && Array.isArray(cleaned.Pasos_Analizados)) {
-            cleaned.Pasos_Analizados = cleaned.Pasos_Analizados.map(paso => {
-                const cleanedPaso = { ...paso };
-
-                // Remove JSON-like formatting from resultado_obtenido_paso_y_estado
-                if (cleanedPaso.resultado_obtenido_paso_y_estado) {
-                    let cleanedResult = cleanedPaso.resultado_obtenido_paso_y_estado;
-
-                    // Remove opening and closing braces if they wrap the entire content
-                    if (cleanedResult.trim().startsWith('{') && cleanedResult.trim().endsWith('}')) {
-                        cleanedResult = cleanedResult.trim().slice(1, -1).trim();
-                    }
-
-                    // Remove JSON property patterns like "estado": "Exitosa", "descripcion": "..."
-                    // and extract just the meaningful text
-                    const jsonPatterns = [
-                        /"estado"\s*:\s*"([^"]+)"\s*,\s*"descripcion"\s*:\s*"([^"]+)"/i,
-                        /"descripcion"\s*:\s*"([^"]+)"\s*,\s*"estado"\s*:\s*"([^"]+)"/i
-                    ];
-
-                    for (const pattern of jsonPatterns) {
-                        const match = cleanedResult.match(pattern);
-                        if (match) {
-                            // Extract estado and descripcion from JSON pattern
-                            const estado = match[1] || match[2];
-                            const descripcion = match[2] || match[1];
-                            cleanedResult = `${estado}: ${descripcion}`;
-                            break;
-                        }
-                    }
-
-                    cleanedPaso.resultado_obtenido_paso_y_estado = cleanedResult.trim();
-                }
-
-                if (cleanedPaso.dato_de_entrada_paso === 'N/A') {
-                    cleanedPaso.dato_de_entrada_paso = '';
-                }
-
-                return cleanedPaso;
-            });
-        }
-
-        return cleaned;
     };
 
     useEffect(() => {
